edit

+++ sbin/spark/split_by_month.sh
@@ -1,19 +1,27 @@
 #!/usr/bin/env bash
 
-home="data/dump/2nd"
+home="data/lineagem"
 
-#nate_entertainment 2017 안됨
-# -filename data/dump/2nd/daum_baseball/2017.json.bz2 -result data/dump/2nd/daum_baseball/2017.by_month
+dry=""
 
-for dir_name in $(hadoop fs -ls -d ${home}/daum_baseball/ | perl -ple 's/\s+/\t/g' | cut -f8-) ; do
-    db_name=$(basename ${dir_name})
+for input in $(\ls -d ${home}/자유.json.bz2) ; do
+    year=$(basename ${input})
+    year="${year/.json.bz2/}"
 
-    for input in $(ls ${home}/${db_name}/2017.json.bz2) ; do
-        year=$(basename ${input})
-        year="${year/.json.bz2/}"
+    output="${input/.json.bz2/.by_month}"
+    if [ -f "${output}/_SUCCESS" ] ; then
+        echo "skip ${input}"
+        continue
+    fi
 
-        output="${input/.json.bz2/.by_month}"
-        echo ${input} ${output}
+    echo ${input} ${output}
+
+    if [ "${dry}" == "" ] ; then
+        # 입력 파일 업로드
+        input_path=$(dirname ${input})
+        hadoop fs -mkdir -p ${input_path}
+        hadoop fs -rm -f -skipTrash ${input}
+        hadoop fs -put ${input} ${input_path}
 
         # 이전 결과 파일 삭제
         hadoop fs -rm -r -f -skipTrash ${output}
@@ -22,8 +30,7 @@ for dir_name in $(hadoop fs -ls -d ${home}/daum_baseball/ | perl -ple 's/\s+/\t/
         spark-submit \
             --master yarn \
             --deploy-mode client \
-            --driver-memory 16g \
-            --num-executors 18 \
+            --num-executors 9 \
             --executor-cores 2 \
             --verbose \
             SparkBatchSplitDate.py -filename ${input} -result ${output}
@@ -31,5 +38,5 @@ for dir_name in $(hadoop fs -ls -d ${home}/daum_baseball/ | perl -ple 's/\s+/\t/
         # 결과 파일 다운로드
         rm -rf ${output} && sync
         hadoop fs -get ${output} ${output}
-    done
+    fi
 done
